requirements.txt
selenium
colorama
python-dotenv
